Move default primary/accent colors out of ha-style
